Berry, Growing Stalagmites, and Rotated Trees!
All stalagmites grow now. Have chance of growing same level as self, or
next smallest level. Break when not on solid surface (as torches).

Trees now match your orientation when you place them. There is an error
with Spheretree wood,  because only 3 orientations are possible, and
there are 3 textures -- so, need 6 orientations. 3 directions do not
work.

Added purple berry and recipe to shrivel green berries, and to condense
green and purple berries.